requirements.txt
discord.py==1.3.4
requests==2.20.0
Pillow==7.2.0